requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.63
boto3-stubs==1.40.63
botocore==1.40.63
botocore-stubs==1.40.63
click==8.3.0
fastapi==0.120.3
h11==0.16.0
idna==3.11
jmespath==1.0.1
mypy-boto3-s3==1.40.61
pillow==12.0.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
qrcode==8.2
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
types-awscrt==0.28.2
types-qrcode==8.2.0.20250914
types-s3transfer==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
